# docs/swagger.yaml
basePath: /
definitions:
  course.courseData:
    properties:
      info:
        type: string
      name:
        type: string
      tid:
        type: integer
    required:
    - name
    - tid
    type: object
  course.studentsData:
    properties:
      accounts:
        items:
          type: string
        type: array
    required:
    - accounts
    type: object
  entity.CourseResp:
    properties:
      cid:
        type: integer
      info:
        type: string
      name:
        type: string
      tbegin:
        type: string
      teacher:
        type: string
      tend:
        type: string
      tid:
        type: integer
      tname:
        type: string
    required:
    - tname
    type: object
  entity.ExperimentReq:
    properties:
      begin:
        description: YYYY-MM-DD HH-MM-SS
        type: string
      end:
        type: string
      info:
        type: string
      name:
        type: string
    type: object
...
  entity.SubmissionResp:
    properties:
      groups:
        items:
          $ref: '#/definitions/entity.UserInfoSimple'
        type: array
      info:
        type: string
      name:
        type: string
      readme:
        type: string
      status:
        type: boolean
      time:
        type: string
      type:
        type: integer
      url:
        type: string
    type: object
...
  entity.UserInfoRes:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: integer
    type: object
  entity.UserInfoSimple:
    properties:
      account:
        type: string
      name:
        type: string
    type: object
  experiment.SubmissionReq:
    properties:
      info:
        type: string
      name:
        type: string
      readme:
        type: string
      type:
        type: integer
      url:
        type: string
    type: object
...
info:
  contact:
    name: Mogg
    url: https://github.com/MarcusMogg
  description: buaashow is a sample RESTful api server.
  title: buaashow
  version: "1.0"
paths:
  /course:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.CourseResp'
            type: array
      summary: 获取与当前用户相关的课程(教师创建、学生加入) 需用户登录
      tags:
      - course
    post:
      consumes:
      - application/json
      parameters:
      - description: 课程信息
        in: body
        name: coursedata
        required: true
        schema:
          $ref: '#/definitions/course.courseData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CourseResp'
      summary: 创建课程 需教师登录
      tags:
      - course
  /course/{cid}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: cid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 删除课程,需用户登录，当前用户需要是课程创建者
      tags:
      - course
  /course/{cid}/exp:
    post:
      parameters:
      - description: Course ID
        in: path
        name: cid
        required: true
        type: integer
      - description: 实验信息
        in: body
        name: exp
        required: true
        schema:
          $ref: '#/definitions/entity.ExperimentReq'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 创建实验 需教师登录
      tags:
      - exp
...
  /course/{cid}/student/{uid}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: cid
        required: true
        type: integer
      - description: Student Account
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 删除学生,需用户登录，当前用户有课程管理权限
      tags:
      - course
  /course/{id}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CourseResp'
      summary: 获取课程信息
      tags:
      - course
  /course/{id}/students:
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.UserInfoRes'
            type: array
      summary: 查看课程学生列表
      tags:
      - course
    post:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: 学生账号
        in: body
        name: accounts
        required: true
        schema:
          $ref: '#/definitions/course.studentsData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course.studentsData'
      summary: 导入学生, 需用户登录，当前用户有课程管理权限
      tags:
      - course
...
  /exp/{id}/file:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 选择上传文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 添加实验资源文件,文件存在的话会被替换
      tags:
      - exp
...
  /exp/{id}/submit:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.SubmissionResp'
      summary: 学生提交信息
      tags:
      - exp
    post:
      parameters:
      - description: Exp ID
        in: path
        name: id
        required: true
        type: integer
      - description: 实验信息
        in: body
        name: exp
        required: true
        schema:
          $ref: '#/definitions/experiment.SubmissionReq'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 提交作业
      tags:
      - exp
...
swagger: "2.0"